Mdahrea/minorupdates (#7)
* Trying out some things; untested

* Rename this.data to this.settings

* One class per file (and named)

* Missing some data->settings

* Missing some data->settings

* Remove _setDeltaHP

* Fix config overrides

* Restore log level
