+++ module.json
@@ -16,7 +16,7 @@
   ],
   "url": "https://github.com/Moerill/token-mold",
   "bugs": "https://github.com/Moerill/token-mold/issues",
-  "version": "2.22.2",
+  "version": "2.22.3",
   "compatibility": {
     "minimum": "12",
     "verified": "12",
@@ -68,5 +68,5 @@
   ],
   "socket": false,
   "manifest": "https://raw.githubusercontent.com/Moerill/token-mold/master/module.json",
-  "download": "https://github.com/Moerill/token-mold/releases/download/v2.22.2/v2.22.2.zip"
+  "download": "https://github.com/Moerill/token-mold/releases/download/v2.22.3/v2.22.3.zip"
 }
