+++ scripts/token-consts.js
@@ -1,0 +1,6 @@
+export default class TokenConsts {
+    static SUPPORTED_SYSTEMS = ["dnd5e", "pf2e", "sfrpg", "sw5e", "dcc"];
+    static SUPPORTED_ROLLHP = ["dnd5e", "sw5e", "dcc"];
+    static SUPPORTED_CREATURESIZE = ["dnd5e", "pf2e"];
+    static SUPPORTED_5ESKILLS = ["dnd5e", "sw5e"];
+}
