+++ templates/overlay.html
@@ -1,6 +1,6 @@
 <form id="{{id}}" class="{{classes}}" data-appid="{{appId}}" autocomplete="off">
-		{{#each stats}}
-			<span class='value'>{{this.value}}</span>
-			<span class='icon fas'>{{{this.icon}}}</span>
-		{{/each}}
+    {{#each stats}}
+    <span class='value'>{{this.value}}</span>
+    <span class='icon fas'>{{{this.icon}}}</span>
+    {{/each}}
 </form>
